[FIX] sustainability_purchase: Fix traceback with partner merging wizard
The native partner merging wizard doesn't support fields of type `reference` because it is never used by Odoo on model `res.partner`.
Because we add a `reference` field on `res.partner`, we got a traceback using this wizard on a record with this field filled. This commit fix it.

C.f. ONS(project.task,115597)